+++ apps/rig_inbound_gateway/test/support/conn_case.ex
@@ -48,7 +48,7 @@ defmodule RigInboundGatewayWeb.ConnCase do
               %{
                 "id" => "get-movies",
                 "method" => "GET",
-                "not_secured" => true,
+                "secured" => false,
                 "path" => "/myapi/movies"
               }
             ]